Misc refactoring
