+++ src/arduino/collision.cpp
@@ -1,0 +1,17 @@
+
+#include "collision.h"
+
+namespace robot {
+
+	Sensor *distance_sensors = nullptr;
+	uint8_t distance_sensor_count = 0;
+
+	void set_distance_sensors(uint8_t count, Sensor *sensors) {
+
+	}
+
+	void check_distance_sensors() {
+
+	}
+
+}
