+++ src/arduino/drive.cpp
@@ -1,0 +1,15 @@
+
+#include "drive.h"
+
+namespace robot {
+	namespace drive {
+
+		int32_t target_left_encoder_value = 0, target_right_encoder_value = 0;
+		bool is_moving = false;
+
+		void update_motor_speeds() {
+
+		}
+		
+	}
+}
